[proliferate] move anyharness local capabilities under adapters
