+++ anyharness/crates/anyharness-lib/src/adapters/files/mod.rs
@@ -1,4 +1,3 @@
-pub mod runtime;
 pub mod safety;
 pub mod service;
 pub mod types;
