+++ anyharness/crates/anyharness-lib/src/adapters/files/safety.rs
