+++ anyharness/crates/anyharness-lib/src/adapters/git/diff.rs
