+++ anyharness/crates/anyharness-lib/src/adapters/git/mod.rs
@@ -3,7 +3,6 @@ mod default_branch;
 mod diff;
 pub mod executor;
 pub mod file_search;
-pub mod mobility_delta;
 pub mod parse_status;
 pub mod service;
 pub mod types;
