+++ anyharness/crates/anyharness-lib/src/adapters/mod.rs
@@ -1,0 +1,4 @@
+pub mod files;
+pub mod git;
+pub mod hosting;
+pub mod processes;
