+++ anyharness/crates/anyharness-lib/src/app/mod.rs
@@ -2,6 +2,8 @@ use std::path::PathBuf;
 use std::sync::Arc;
 
 use crate::acp::manager::AcpManager;
+use crate::adapters::git::WorkspaceFileSearchCache;
+use crate::adapters::processes::ProcessService;
 use crate::agents::catalog::{LaunchCatalogService, ModelCatalogService};
 use crate::agents::reconcile_execution::AgentReconcileService;
 use crate::agents::seed::AgentSeedStore;
@@ -11,15 +13,12 @@ use crate::cowork::mcp_auth::CoworkMcpAuth;
 use crate::cowork::runtime::{CoworkRuntime, CoworkSessionHooks};
 use crate::cowork::service::CoworkService;
 use crate::cowork::store::CoworkStore;
-use crate::files::runtime::WorkspaceFilesRuntime;
-use crate::git::WorkspaceFileSearchCache;
 use crate::mobility::service::MobilityService;
 use crate::mobility::store::MobilityStore;
 use crate::persistence::Db;
 use crate::plans::runtime::PlanRuntime;
 use crate::plans::service::PlanService;
 use crate::plans::store::PlanStore;
-use crate::processes::ProcessService;
 use crate::repo_roots::service::RepoRootService;
 use crate::repo_roots::store::RepoRootStore;
 use crate::reviews::hooks::ReviewSessionHooks;
@@ -46,6 +45,7 @@ use crate::terminals::TerminalService;
 use crate::workspaces::access_gate::WorkspaceAccessGate;
 use crate::workspaces::access_store::WorkspaceAccessStore;
 use crate::workspaces::checkout_gate::CheckoutDeletionGate;
+use crate::workspaces::files_runtime::WorkspaceFilesRuntime;
 use crate::workspaces::inventory::WorktreeInventoryService;
 use crate::workspaces::operation_gate::WorkspaceOperationGate;
 use crate::workspaces::purge::WorkspacePurgeService;
