+++ anyharness/crates/anyharness-lib/src/lib.rs
@@ -1,16 +1,13 @@
 pub mod acp;
+pub mod adapters;
 pub mod agents;
 pub mod api;
 pub mod app;
 pub mod cowork;
-pub mod files;
-pub mod git;
-pub mod hosting;
 pub mod mobility;
 pub mod origin;
 pub mod persistence;
 pub mod plans;
-pub mod processes;
 pub mod repo_roots;
 pub mod reviews;
 pub mod sessions;
