+++ anyharness/crates/anyharness-lib/src/mobility/mod.rs
@@ -1,3 +1,4 @@
 pub mod model;
 pub mod service;
 pub mod store;
+pub mod workspace_delta;
